refactor(cli): split cmd_project.cpp by sub-command (P11-1)
Roadmap item P11-1. cmd_project.cpp was a single ~1,174-line
translation unit holding the pin/bump, undo, and unpin handlers
plus the shared pin-rewrite helpers they all depend on. Following
the established R2-4 extraction pattern, split it into focused
sibling TUs behind a PRIVATE internal header:

  cmd_project.cpp          dispatcher (cmd_project) + do_unpin   156 lines
  cmd_project_internal.hpp private entry-point surface            85 lines
  cmd_project_common.cpp   shared pin-file / project-root helpers 344 lines
  cmd_project_bump.cpp     do_bump (pin / bump)                  540 lines
  cmd_project_undo.cpp     do_undo                               189 lines

cmd_project.cpp drops from 1174 to 156 lines. Extracted code is
byte-identical (only namespace-wrapped into pulp_cli::project_detail
and re-headered) — pure mechanical move, zero behavior change. The
new .cpp files are registered next to cmd_project.cpp in
tools/cli/CMakeLists.txt; cmd_project_internal.hpp lives next to the
.cpp files under tools/cli/, not in any public include tree.

test_cli_project_command.cpp includes the cmd_project family
directly to exercise anonymous-namespace helpers; updated to include
all four split TUs and pull in the project_detail namespace so the
existing unqualified test code keeps resolving — assertions unchanged.

Build: pulp-cli (pulp-cpp) clean, exit 0.
Tests: pulp-test-cli-project-command 35/35, pulp-test-cli-project-bump
24/24, pulp-test-cli-project-bump-all 4/4; all 20 cmd_project ctest
cases + all 33 cli-* ctest entries pass.
